+++ web/src/App.jsx
@@ -1,4 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { AuthProvider } from './hooks/useAuth.jsx';
 import Header from './components/Header';
 import Home from './pages/Home';
 import PluginDetail from './pages/PluginDetail';
@@ -6,18 +7,20 @@ import Developer from './pages/Developer';
 
 function App() {
   return (
-    <BrowserRouter>
-      <div className="min-h-screen bg-white">
-        <Header />
-        <main>
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/plugin/:id" element={<PluginDetail />} />
-            <Route path="/developer" element={<Developer />} />
-          </Routes>
-        </main>
-      </div>
-    </BrowserRouter>
+    <AuthProvider>
+      <BrowserRouter>
+        <div className="min-h-screen bg-white">
+          <Header />
+          <main>
+            <Routes>
+              <Route path="/" element={<Home />} />
+              <Route path="/plugin/:id" element={<PluginDetail />} />
+              <Route path="/developer" element={<Developer />} />
+            </Routes>
+          </main>
+        </div>
+      </BrowserRouter>
+    </AuthProvider>
   );
 }